Moved help strings into language yaml file
