+++ util.py
@@ -14,6 +14,7 @@ from discord.ext.commands import (
     BadArgument,
     CommandError,
     Command,
+    Group,
 )
 
 from database import run_in_thread
@@ -189,3 +190,59 @@ async def read_complete_message(message: Message) -> Tuple[str, List[File], Opti
         embed = None
 
     return message.content, [await attachment_to_file(attachment) for attachment in message.attachments], embed
+
+
+async def send_help(ctx: Context, command_name: Optional[Union[str, Command]]):
+    def format_command(cmd: Command) -> str:
+        doc = " - " + cmd.short_doc if cmd.short_doc else ""
+        return f"`{cmd.name}`{doc}"
+
+    async def add_commands(cog_name: str, commands: List[Command]):
+        desc: List[str] = []
+        for cmd in commands:
+            if not cmd.hidden and await can_run_command(cmd, ctx):
+                desc.append(format_command(cmd))
+        if desc:
+            embed.add_field(name=cog_name, value="\n".join(desc), inline=False)
+
+    prefix: str = await get_prefix()
+    embed = Embed(title=translations.help, color=0x008080)
+    if command_name is None:
+        for cog in ctx.bot.cogs.values():
+            await add_commands(cog.qualified_name, cog.get_commands())
+        await add_commands(translations.no_category, [command for command in ctx.bot.commands if command.cog is None])
+
+        embed.add_field(name="** **", value=translations.f_help_usage(prefix), inline=False)
+
+        await ctx.send(embed=embed)
+        return
+
+    if isinstance(command_name, str):
+        cog: Optional[Cog] = ctx.bot.get_cog(command_name)
+        if cog is not None:
+            await add_commands(cog.qualified_name, cog.get_commands())
+            await ctx.send(embed=embed)
+            return
+
+        command: Optional[Union[Command, Group]] = ctx.bot.get_command(command_name)
+        if command is None:
+            raise CommandError(translations.cog_or_command_not_found)
+    else:
+        command: Command = command_name
+
+    description = prefix
+    if command.full_parent_name:
+        description += command.full_parent_name + " "
+    if command.aliases:
+        description += "[" + "|".join([command.name] + command.aliases) + "] "
+    else:
+        description += command.name + " "
+    description += command.signature
+
+    embed.description = f"```css\n{description.strip()}\n```"
+    embed.add_field(name=translations.description, value=command.help, inline=False)
+
+    if isinstance(command, Group):
+        await add_commands(translations.subcommands, command.commands)
+
+    await ctx.send(embed=embed)
